requirements.txt
asgiref==3.3.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.6
django-cors-headers==3.8.0
django-rest-auth==0.9.5
django-rest-framework-base64-fields==0.1.3
djangorestframework==3.12.4
drf-extra-fields==3.1.1
drf-yasg==1.20.0
idna==3.2
importlib-metadata==4.8.1
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.0
pika==1.2.0
plumbum==1.7.0
psycopg2==2.8.6
PyJWT==2.3.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
rpyc==5.0.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
serpent==1.40
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1